[MOVE] : post 패키지 위치 domain 안으로 수정
